Refactor docs
