+++ src/index.ts
@@ -1,2 +1,2 @@
 export { defineField } from './defineField';
-export type { InferFieldShape, InferFieldSchema } from './core/types';
+export type { InferFieldShape, InferFieldSchema } from './types';
